Register in the client context a state to avoid reconnecting on Azure.
This commit refactors the code (mostly moves it) to ease readability. It
also add an Azure context that will be kept between the files access in
a query. Previously if you were using a credential_chain provider
when you query multiple files for each of them, we will have initiate a
new connection and identify at the Azure AD (now Entra). Now this will
be only performed once by query.
