+++ CMakeLists.txt
@@ -10,7 +10,9 @@ include_directories(src/include)
 set(CMAKE_CXX_STANDARD 14)
 set(CMAKE_CXX_STANDARD_REQUIRED True)
 
-set(EXTENSION_SOURCES src/azure_extension.cpp src/azure_secret.cpp)
+set(EXTENSION_SOURCES
+    src/azure_extension.cpp src/azure_secret.cpp
+    src/azure_storage_account_client.cpp src/azure_filesystem.cpp)
 add_library(${EXTENSION_NAME} STATIC ${EXTENSION_SOURCES})
 
 set(PARAMETERS "-warnings")
