+++ src/include/azure_storage_account_client.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "duckdb/common/file_opener.hpp"
+#include <azure/storage/blobs/blob_service_client.hpp>
+#include <string>
+
+namespace duckdb {
+
+Azure::Storage::Blobs::BlobServiceClient ConnectToStorageAccount(FileOpener *opener, const std::string &path);
+
+} // namespace duckdb
